update nhom 1

+++ src/quanlysinhvien/controller/DanhMucHPController.java
@@ -10,10 +10,8 @@ import java.io.IOException;
 import java.util.ArrayList;
 
 import javax.swing.JComboBox;
-import javax.swing.JOptionPane;
 import javax.swing.JTable;
 import javax.swing.JTextField;
-import javax.swing.table.DefaultTableModel;
 
 import org.apache.poi.ss.usermodel.Cell;
 import org.apache.poi.ss.usermodel.Row;
@@ -30,14 +28,14 @@ public class DanhMucHPController {
 	private JTable table;
 	private PanelDanhMucHP danhMucHocPhan;
 	private ArrayList<HocPhan> dsHocPhan;
-	
+
 	public DanhMucHPController(PanelDanhMucHP danhMucHocPhan) {
 		this.danhMucHocPhan = danhMucHocPhan;
 		this.table = danhMucHocPhan.getTable();
 		this.tfTimIdHP = danhMucHocPhan.getTfTimIdHP();
 		this.tfTimTenHP = danhMucHocPhan.getTfTimTenHP();
 		this.khoaVienCB = danhMucHocPhan.getKhoaVienCB();
-		
+
 		try {
 			dsHocPhan = readFile();
 		} catch (IOException e) {
@@ -46,17 +44,17 @@ public class DanhMucHPController {
 			System.out.println("Error danhMucHP: " + e);
 		}
 		this.danhMucHocPhan.loadData(table, dsHocPhan, "All", "");
-		
+
 		setAction();
 	}
-	
+
 	private void setAction() {
 		khoaVienCB.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				danhMucHocPhan.loadData(table, dsHocPhan, (String) khoaVienCB.getSelectedItem(), "");
 			}
 		});
-		
+
 		tfTimIdHP.addKeyListener(new KeyAdapter() {
 			@Override
 			public void keyPressed(KeyEvent e) {
@@ -65,7 +63,7 @@ public class DanhMucHPController {
 				}
 			}
 		});
-		
+
 		tfTimTenHP.addKeyListener(new KeyAdapter() {
 			@Override
 			public void keyPressed(KeyEvent e) {
@@ -75,49 +73,49 @@ public class DanhMucHPController {
 			}
 		});
 	}
-	
+
 	private ArrayList<HocPhan> readFile() throws IOException {
 		dsHocPhan = new ArrayList<HocPhan>();
-		
+
 		File file = new File("quanlysinhvien/danhsachhocphan/dsHocphan.xlsx");
 		FileInputStream inputStream = new FileInputStream(file);
-		
+
 		Workbook workbook = new XSSFWorkbook(inputStream);
 		Sheet sheet = workbook.getSheetAt(0);
-		
+
 		int rowCount = sheet.getLastRowNum();
-		
+
 		for (int i = 1; i <= rowCount; i++) {
-			
+
 			HocPhan hp = new HocPhan();
-			
+
 			Row row = sheet.getRow(i);
 			Cell cell;
-			
+
 			cell = row.getCell(1);
 			hp.setIdHocPhan(cell.getStringCellValue());
-			
+
 			cell = row.getCell(2);
 			hp.setTenHP(cell.getStringCellValue());
-			
+
 			cell = row.getCell(3);
 			hp.setSoTinChi((int) cell.getNumericCellValue());
-			
+
 			cell = row.getCell(4);
-			hp.setSoTCHocPhi((int) cell.getNumericCellValue());
-			
+			hp.setSoTCHocPhi(cell.getNumericCellValue());
+
 			cell = row.getCell(5);
 			hp.setIdNganh(cell.getStringCellValue());
-			
+
 			cell = row.getCell(6);
 			hp.setTrongSo((double) cell.getNumericCellValue());
-			
+
 			dsHocPhan.add(hp);
 		}
 		workbook.close();
 		inputStream.close();
-		
+
 		return dsHocPhan;
 	}
-	
+
 }
